Admins bypass filter

+++ src/Events/message/message.js
@@ -10,7 +10,7 @@ module.exports = class extends Event {
 
     // FILTER
     const file = f.loadJSON(`${message.guild.id}.json`);
-    if (file) {
+    if (file && !message.member.hasPermission("ADMINISTRATOR")) {
       file.phrases.forEach(phrase => {
         if (message.content.toLowerCase().includes(phrase)) {
           message.delete();
